Refactor existing code into Project component
- Move existing code to Project component
- Create necessary modular sass files for different components, as well as modular media sass files
- Finish creating routes on "PageLayout" and "App" files
