+++ src/App.scss
@@ -13,133 +13,3 @@ body {
     padding: 0;
     font-family: 'Neucha', cursive;
 }
-
-
-// App content
-.App {
-    //width: 100vw;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-    background-color: #2f74c0;
-    gap: 2rem;
-}
-
-
-// Main content
-.main_content {
-    display: flex;
-    justify-content: space-between;
-    width: 100%;
-    height: 60%;
-    padding: 0 5%;
-    gap: 2rem;
-}
-
-
-.active_tab {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    flex: 0 0 48%;
-    height: fit-content;
-    max-height: 100%;
-    gap: 1rem;
-    padding: 1rem;
-    background-color: rgb(77, 199, 221);
-
-    .tab_top_row {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-
-        h3 {
-            margin: 0;
-        }
-
-        .tab_top_row_search_input_container {
-
-            input {
-               border-radius: .25rem;
-               padding: .2rem .5rem;
-            }
-       }
-
-        .tab_top_row_sort_buttons {
-            display: flex;
-            align-items: center;
-            justify-content: flex-end;
-            gap: .75rem;
-
-            .sort_button_container {
-                background-color: grey;
-                i {
-                    color: black;
-                    cursor:pointer;
-                }
-            }  
-        }
-    }
-
-    overflow-y: auto;
-
-    &.dragging_over {
-        background-color: rgb(54, 146, 162);
-    }
-}
-
-.completed_tab {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    flex: 0 0 48%;
-    height: fit-content;
-    max-height: 100%;
-    gap: 1rem;
-    padding: 1rem;
-    background-color: rgb(255, 106, 0);
-    transition: all .9s ease-in-out;
-
-    .tab_top_row {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-
-        h3 {
-            margin: 0;
-        }
-
-        .tab_top_row_search_input_container {
-
-             input {
-                border-radius: .25rem;
-                padding: .2rem .5rem;
-             }
-        }
-
-        .tab_top_row_sort_buttons {
-            display: flex;
-            align-items: center;
-            justify-content: flex-end;
-            gap: .75rem;
-
-            .sort_button_container {
-                background-color: grey;
-                i {
-                    color: black;
-                    cursor:pointer;
-                }
-            }  
-        }
-    }
-
-    overflow-y: auto;
-
-    &.dragging_over {
-        background-color: rgb(174, 75, 5);
-    }
-}
-
-@import './media';
