--- src/components/InputField/InputField.module.scss
@@ -1,51 +1,0 @@
-// InputField content
-.input {
-    display: flex;
-    width: 90%;
-    position: relative;
-    align-items: center;
-
-    .input_box {
-        width: 100%;
-        border-radius: 50px;
-        padding: 20px 30px;
-        font-size: 25px;
-        border: none;
-        box-shadow: inset 0 0 5px black;
-        transition: all .2s ease;
-
-        &:focus {
-            outline: none;
-            box-shadow: 0 0 10px 1000px rgba(0, 0, 0, .5);
-        }
-    }
-
-    .input_submit {
-        position: absolute;
-        top: 0%;
-        right: 0%;
-        padding: 1rem;
-        width: fit-content;
-        height: fit-content;
-        border-radius: 50px;
-        border: none;
-        background-color: #2f74c0;
-        color: white;
-        font-size: 15px;
-        width: 50px;
-        height: 50px;
-        margin: 12px;
-        cursor: pointer;
-        box-shadow: 0 0 10px black;
-        transition: background-color .2s ease-in-out, scale .2s ease;
-
-        &:hover {
-            background-color: #388ae2;
-        }
-
-        &:active {
-            scale: .8;
-            box-shadow: 0 0 5px black;
-        }
-    }
-}
