+++ src/layouts/PageLayout.tsx
@@ -1,7 +1,17 @@
 import React from 'react'
+import { Outlet } from 'react-router-dom';
+
+import Footer from './Footer';
+import Navbar from './Navbar';
+import styles from './Layouts.module.scss';
+
 
 export default function PageLayout() {
   return (
-    <div>PageLayout</div>
-  )
+      <div className={styles.app_container}>
+          <Navbar />
+          <Outlet />
+          <Footer />
+      </div>
+  );
 }
