+++ src/layouts/_media.scss
@@ -1,6 +1,6 @@
 // Media query for mobile content (portrait)
 @media (max-width: 500px) {
-    .App {
+    .app_container {
         gap: 1rem;
 
         .heading {
