+++ src/pages/Project/Project.module.scss
@@ -1,0 +1,178 @@
+// InputField content
+.input {
+    display: flex;
+    width: 90%;
+    position: relative;
+    align-items: center;
+
+    .input_box {
+        width: 100%;
+        border-radius: 50px;
+        padding: 20px 30px;
+        font-size: 25px;
+        border: none;
+        box-shadow: inset 0 0 5px black;
+        transition: all .2s ease;
+
+        &:focus {
+            outline: none;
+            box-shadow: 0 0 10px 1000px rgba(0, 0, 0, .5);
+        }
+    }
+
+    .input_submit {
+        position: absolute;
+        top: 0%;
+        right: 0%;
+        padding: 1rem;
+        width: fit-content;
+        height: fit-content;
+        border-radius: 50px;
+        border: none;
+        background-color: #2f74c0;
+        color: white;
+        font-size: 15px;
+        width: 50px;
+        height: 50px;
+        margin: 12px;
+        cursor: pointer;
+        box-shadow: 0 0 10px black;
+        transition: background-color .2s ease-in-out, scale .2s ease;
+
+        &:hover {
+            background-color: #388ae2;
+        }
+
+        &:active {
+            scale: .8;
+            box-shadow: 0 0 5px black;
+        }
+    }
+}
+
+
+.project_container {
+    //width: 100vw;
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    background-color: #2f74c0;
+    gap: 2rem;
+}
+
+
+// Main content
+.main_content {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    height: 60%;
+    padding: 0 5%;
+    gap: 2rem;
+}
+
+
+.active_tab {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    flex: 0 0 48%;
+    height: fit-content;
+    max-height: 100%;
+    gap: 1rem;
+    padding: 1rem;
+    background-color: rgb(77, 199, 221);
+
+    .tab_top_row {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+
+        h3 {
+            margin: 0;
+        }
+
+        .tab_top_row_search_input_container {
+
+            input {
+               border-radius: .25rem;
+               padding: .2rem .5rem;
+            }
+       }
+
+        .tab_top_row_sort_buttons {
+            display: flex;
+            align-items: center;
+            justify-content: flex-end;
+            gap: .75rem;
+
+            .sort_button_container {
+                background-color: grey;
+                i {
+                    color: black;
+                    cursor:pointer;
+                }
+            }  
+        }
+    }
+
+    overflow-y: auto;
+
+    &.dragging_over {
+        background-color: rgb(54, 146, 162);
+    }
+}
+
+.completed_tab {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    flex: 0 0 48%;
+    height: fit-content;
+    max-height: 100%;
+    gap: 1rem;
+    padding: 1rem;
+    background-color: rgb(255, 106, 0);
+    transition: all .9s ease-in-out;
+
+    .tab_top_row {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+
+        h3 {
+            margin: 0;
+        }
+
+        .tab_top_row_search_input_container {
+
+             input {
+                border-radius: .25rem;
+                padding: .2rem .5rem;
+             }
+        }
+
+        .tab_top_row_sort_buttons {
+            display: flex;
+            align-items: center;
+            justify-content: flex-end;
+            gap: .75rem;
+
+            .sort_button_container {
+                background-color: grey;
+                i {
+                    color: black;
+                    cursor:pointer;
+                }
+            }  
+        }
+    }
+
+    overflow-y: auto;
+
+    &.dragging_over {
+        background-color: rgb(174, 75, 5);
+    }
+}
